@@ -11,32 +11,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.2.0",
-    "@angular/common": "~10.2.0",
-    "@angular/compiler": "~10.2.0",
-    "@angular/core": "~10.2.0",
-    "@angular/forms": "~10.2.0",
-    "@angular/platform-browser": "~10.2.0",
-    "@angular/platform-browser-dynamic": "~10.2.0",
-    "@angular/router": "~10.2.0",
+    "@angular/animations": "^11.2.0",
+    "@angular/cdk": "^11.2.0",
+    "@angular/common": "~11.2.0",
+    "@angular/compiler": "~11.2.0",
+    "@angular/core": "~11.2.0",
+    "@angular/forms": "~11.2.0",
+    "@angular/material": "^11.2.0",
+    "@angular/platform-browser": "~11.2.0",
+    "@angular/platform-browser-dynamic": "~11.2.0",
+    "@angular/router": "~11.2.0",
+    "@igniteui/material-icons-extended": "^2.4.0",
+    "@material/card": "^10.0.0",
     "bootstrap": "^4.6.0",
-    "eslint": "^7.19.0",
+    "eslint": "^7.20.0",
+    "hammerjs": "^2.0.8",
+    "igniteui-angular": "^10.2.13",
     "jquery": "^3.5.1",
+    "jszip": "^3.6.0",
+    "minireset.css": "~0.0.4",
+    "resize-observer-polyfill": "^1.5.1",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
+    "web-animations-js": "^2.3.2",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1002.0",
-    "@angular/cli": "~10.2.0",
-    "@angular/compiler-cli": "~10.2.0",
-    "@types/node": "^12.11.1",
-    "@types/jasmine": "~3.5.0",
+    "@angular-builders/custom-webpack": "^11.0.0",
+    "@angular-devkit/build-angular": "~0.1102.0",
+    "@angular/cli": "^11.2.0",
+    "@angular/compiler-cli": "~11.2.0",
+    "@igniteui/angular-schematics": "~10.2.620",
+    "@types/hammerjs": "^2.0.36",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^12.20.1",
+    "angular-cli": "^1.0.0-beta.28.3",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.1.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
@@ -44,6 +58,7 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.0.2"
+    "typescript": "^4.1.5",
+    "webpack": "^5.21.2"
   }
 }
